Finish file system implementation and tests
