+++ IntellitectTerminal.Tests/Tests/CommandServiceTests.cs
@@ -24,14 +24,14 @@ namespace IntellitectTerminal.Tests
         public void Request_ExistingUser_ReturnsLevelOneQuestion()
         {
             TestData.AddNewChallenges();
-            Assert.Equal(1, UnderTest.Request(TestData.AddUser().UserId).Level);
+            Assert.Equal(1, UnderTest.Request(TestData.AddUserWithOnlyGuid().UserId).Level);
         }
 
         [Fact]
         public void Request_ExistingUserWithLevelOneSubmission_ReturnsLevelTwoQuestion()
         {
             Challenge challenge = TestData.AddNewChallenges().Where(x=>x.Level == 1).First();
-            User user = TestData.AddUser();
+            User user = TestData.AddUserWithOnlyGuid();
             TestData.AddSubmission(user, challenge, true);
             Assert.Equal(2, UnderTest.Request(user.UserId).Level);
         }
